Update Names for Bundles and Add Error Codes when Parameters are Missing.
We updated the names of the scripts to make it clear when multiple Jobs are bundled into a single script.

We also added exit code and error messages if a required parameter is missing. The exit code will be 610 and the missing parameter will be included in the error message.