Added retry logic for the 500 error.
We saw occasional 500 errors when polling to check the Job Completion Status after the Job was added. We added retry logic if that 500 error presents itself. Now that hte retry logic is in place, the 500 error would need to exist 5 times in a row over 30 seconds before the Job will fail.